[CHORE] CGFloat extension 파일로 분리(#196)
