+++ EarthValley80/EarthValley80/Screen/SelectInterest/Components/BubbleScene.swift
@@ -7,17 +7,6 @@
 
 import SpriteKit
 
-extension CGFloat {
-
-    public static func random() -> CGFloat {
-        return (CGFloat(arc4random()) / CGFloat(UInt32.max))
-    }
-
-    public static func random(min: CGFloat, max: CGFloat) -> CGFloat {
-        return CGFloat.random() * (max - min) + min
-    }
-}
-
 final class BubblesScene: SIFloatingCollectionScene {
     var topOffset: CGFloat = 0
